Addressed more comments from meeting, cleanup of dead code after it
