+++ arangod/RocksDBEngine/RocksDBCollection.cpp
@@ -34,7 +34,6 @@
 #include "Basics/VelocyPackHelper.h"
 #include "Basics/debugging.h"
 #include "Cache/BinaryKeyHasher.h"
-#include "Statistics/ServerStatistics.h"
 #include "Cache/Common.h"
 #include "Cache/Manager.h"
 #include "Cache/TransactionalCache.h"
@@ -212,7 +211,7 @@ struct TimeTracker {
   TimeTracker(TimeTracker const&) = delete;
   TimeTracker& operator=(TimeTracker const&) = delete;
 
-  using Metrics = TransactionStatistics::ReadWriteMetrics;
+  using Metrics = RocksDBReadWriteMetrics;
   using Count = void (*)(Metrics& metrics, float time) noexcept;
 
   explicit TimeTracker(std::optional<Metrics>& metrics, Count count) noexcept
@@ -311,16 +310,16 @@ namespace arangodb {
 // from IResearchKludge
 void syncIndexOnCreate(Index&);
 
-RocksDBCollection::RocksDBCollection(LogicalCollection& collection,
-                                     velocypack::Slice info,
-                                     cache::Manager* cacheManager,
-                                     TransactionStatistics& statistics)
+RocksDBCollection::RocksDBCollection(
+    LogicalCollection& collection, velocypack::Slice info,
+    cache::Manager* cacheManager,
+    std::optional<RocksDBReadWriteMetrics>& readWriteMetrics)
     : RocksDBMetaCollection(collection, info),
       _primaryIndex(nullptr),
       _cacheManager(cacheManager),
       _maxCacheValueSize(
           _cacheManager == nullptr ? 0 : _cacheManager->maxCacheValueSize()),
-      _statistics(statistics),
+      _readWriteMetrics(readWriteMetrics),
       _cacheEnabled(_cacheManager != nullptr && !collection.system() &&
                     !collection.isAStub() &&
                     !ServerState::instance()->isCoordinator() &&
@@ -780,9 +779,10 @@ Result RocksDBCollection::truncate(transaction::Methods& trx,
                                    OperationOptions& options,
                                    bool& usedRangeDelete) {
   ::TruncateTimeTracker timeTracker(
-      _statistics._readWriteMetrics, options,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_truncate_sec.count(time); });
+      _readWriteMetrics, options,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_truncate_sec.count(time);
+      });
 
   auto state = RocksDBTransactionState::toState(&trx);
   TRI_ASSERT(!state->isReadOnlyTransaction());
@@ -1075,9 +1075,10 @@ Result RocksDBCollection::lookup(transaction::Methods* trx,
   TRI_IF_FAILURE("LogicalCollection::read") { return Result(TRI_ERROR_DEBUG); }
 
   ::ReadTimeTracker timeTracker(
-      _statistics._readWriteMetrics,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_read_sec.count(time); });
+      _readWriteMetrics,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_read_sec.count(time);
+      });
 
   rocksdb::PinnableSlice ps;
   Result res;
@@ -1121,9 +1122,10 @@ Result RocksDBCollection::lookup(transaction::Methods* trx,
                                  LookupOptions options,
                                  StorageSnapshot const* snapshot) const {
   ::ReadTimeTracker timeTracker(
-      _statistics._readWriteMetrics,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_read_sec.count(time); });
+      _readWriteMetrics,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_read_sec.count(time);
+      });
 
   if (!token.isSet()) {
     return Result{TRI_ERROR_ARANGO_DOCUMENT_KEY_BAD,
@@ -1138,9 +1140,10 @@ Result RocksDBCollection::lookup(transaction::Methods* trx,
                                  MultiDocumentCallback const& cb,
                                  LookupOptions options) const {
   ::ReadTimeTracker timeTracker(
-      _statistics._readWriteMetrics,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_read_sec.count(time); });
+      _readWriteMetrics,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_read_sec.count(time);
+      });
 
   RocksDBMethods* mthd =
       RocksDBTransactionState::toMethods(trx, _logicalCollection.id());
@@ -1205,9 +1208,10 @@ Result RocksDBCollection::insert(transaction::Methods& trx,
                                  velocypack::Slice newDocument,
                                  OperationOptions const& options) {
   ::WriteTimeTracker timeTracker(
-      _statistics._readWriteMetrics, options,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_insert_sec.count(time); });
+      _readWriteMetrics, options,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_insert_sec.count(time);
+      });
 
   TRI_ASSERT(newRevisionId.isSet());
   TRI_ASSERT(newDocument.isObject());
@@ -1235,9 +1239,10 @@ Result RocksDBCollection::update(
     velocypack::Slice previousDocument, RevisionId newRevisionId,
     velocypack::Slice newDocument, OperationOptions const& options) {
   ::WriteTimeTracker timeTracker(
-      _statistics._readWriteMetrics, options,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_update_sec.count(time); });
+      _readWriteMetrics, options,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_update_sec.count(time);
+      });
 
   return performUpdateOrReplace(trx, indexesSnapshot, previousDocumentId,
                                 previousRevisionId, previousDocument,
@@ -1251,9 +1256,10 @@ Result RocksDBCollection::replace(
     velocypack::Slice previousDocument, RevisionId newRevisionId,
     velocypack::Slice newDocument, OperationOptions const& options) {
   ::WriteTimeTracker timeTracker(
-      _statistics._readWriteMetrics, options,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_replace_sec.count(time); });
+      _readWriteMetrics, options,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_replace_sec.count(time);
+      });
 
   return performUpdateOrReplace(trx, indexesSnapshot, previousDocumentId,
                                 previousRevisionId, previousDocument,
@@ -1296,9 +1302,10 @@ Result RocksDBCollection::remove(transaction::Methods& trx,
                                  velocypack::Slice previousDocument,
                                  OperationOptions const& options) {
   ::WriteTimeTracker timeTracker(
-      _statistics._readWriteMetrics, options,
-      [](TransactionStatistics::ReadWriteMetrics& metrics,
-         float time) noexcept { metrics.rocksdb_remove_sec.count(time); });
+      _readWriteMetrics, options,
+      [](RocksDBReadWriteMetrics& metrics, float time) noexcept {
+        metrics.rocksdb_remove_sec.count(time);
+      });
 
   TRI_ASSERT(previousDocumentId.isSet());
   TRI_ASSERT(previousDocument.isObject());
