+++ arangod/Statistics/ServerStatistics.cpp
@@ -27,15 +27,10 @@
 #include "Metrics/CounterBuilder.h"
 #include "Metrics/GaugeBuilder.h"
 #include "Metrics/HistogramBuilder.h"
-#include "Metrics/LogScale.h"
 #include "Metrics/IRegistry.h"
+#include "Metrics/TimeScale.h"
 
 using namespace arangodb;
-
-template<typename T = float>
-struct TimeScale {
-  static metrics::LogScale<T> scale() { return {10., 0.0, 1000.0, 11}; }
-};
 
 DECLARE_COUNTER(arangodb_collection_lock_acquisition_micros_total,
                 "Total amount of collection lock acquisition time [μs]");
@@ -67,31 +62,6 @@ DECLARE_COUNTER(arangodb_read_transactions_total,
 DECLARE_COUNTER(arangodb_dirty_read_transactions_total,
                 "Number of read transactions which can do dirty reads");
 
-DECLARE_COUNTER(arangodb_collection_truncates_total,
-                "Total number of collection truncate operations (excl. "
-                "synchronous replication)");
-DECLARE_COUNTER(arangodb_collection_truncates_replication_total,
-                "Total number of collection truncate operations by synchronous "
-                "replication");
-DECLARE_COUNTER(arangodb_document_writes_total,
-                "Total number of document write operations (excl. synchronous "
-                "replication)");
-DECLARE_COUNTER(
-    arangodb_document_writes_replication_total,
-    "Total number of document write operations by synchronous replication");
-DECLARE_HISTOGRAM(arangodb_document_read_time, TimeScale<>,
-                  "Total time spent in document read operations [s]");
-DECLARE_HISTOGRAM(arangodb_document_insert_time, TimeScale<>,
-                  "Total time spent in document insert operations [s]");
-DECLARE_HISTOGRAM(arangodb_document_replace_time, TimeScale<>,
-                  "Total time spent in document replace operations [s]");
-DECLARE_HISTOGRAM(arangodb_document_remove_time, TimeScale<>,
-                  "Total time spent in document remove operations [s]");
-DECLARE_HISTOGRAM(arangodb_document_update_time, TimeScale<>,
-                  "Total time spent in document update operations [s]");
-DECLARE_HISTOGRAM(arangodb_collection_truncate_time, TimeScale<>,
-                  "Total time spent in collection truncate operations [s]");
-
 TransactionStatistics::TransactionStatistics(metrics::IRegistry& metrics)
     : _metrics(metrics),
       _restTransactionsMemoryUsage(
@@ -115,26 +85,6 @@ TransactionStatistics::TransactionStatistics(metrics::IRegistry& metrics)
       _lockTimes(_metrics.add(arangodb_collection_lock_acquisition_time{})),
       _sequentialLocks(
           _metrics.add(arangodb_collection_lock_sequential_mode_total{})) {}
-
-void TransactionStatistics::setupDocumentMetrics() {
-  if (_readWriteMetrics.has_value()) {
-    return;
-  }
-  // the following metrics are conditional, so we don't initialize them in the
-  // constructor
-  _readWriteMetrics.emplace(ReadWriteMetrics{
-      _metrics.add(arangodb_document_writes_total{}),
-      _metrics.add(arangodb_document_writes_replication_total{}),
-      _metrics.add(arangodb_collection_truncates_total{}),
-      _metrics.add(arangodb_collection_truncates_replication_total{}),
-      _metrics.add(arangodb_document_read_time{}),
-      _metrics.add(arangodb_document_insert_time{}),
-      _metrics.add(arangodb_document_replace_time{}),
-      _metrics.add(arangodb_document_remove_time{}),
-      _metrics.add(arangodb_document_update_time{}),
-      _metrics.add(arangodb_collection_truncate_time{}),
-  });
-}
 
 double ServerStatistics::uptime() const noexcept {
   return StatisticsFeature::time() - _startTime;
