+++ arangod/StorageEngine/StorageEngine.h
@@ -27,6 +27,8 @@
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Basics/Result.h"
 #include "Indexes/IndexFactory.h"
+#include "Metrics/Fwd.h"
+#include "Statistics/ServerStatistics.h"
 #include "StorageEngine/HealthData.h"
 #include "Transaction/ManagerFeature.h"
 #include "Transaction/OperationOrigin.h"
@@ -40,8 +42,6 @@
 #include <vector>
 
 namespace arangodb {
-
-struct TransactionStatistics;
 
 namespace velocypack {
 class Slice;
@@ -378,11 +378,11 @@ class StorageEngine : public application_features::ApplicationFeature {
   virtual bool autoRefillIndexCachesOnFollowers() const = 0;
   virtual void syncIndexCaches();
 
-  virtual TransactionStatistics& transactionStatistics() noexcept = 0;
-  virtual TransactionStatistics const& transactionStatistics()
-      const noexcept = 0;
+  TransactionStatistics& transactionStatistics() noexcept;
+  TransactionStatistics const& transactionStatistics() const noexcept;
 
  protected:
+  void initTransactionStatistics(metrics::IRegistry& metrics);
   void registerCollection(
       TRI_vocbase_t& vocbase,
       std::shared_ptr<arangodb::LogicalCollection> const& collection);
@@ -397,6 +397,7 @@ class StorageEngine : public application_features::ApplicationFeature {
  private:
   std::unique_ptr<IndexFactory> const _indexFactory;
   std::string_view _typeName;
+  std::unique_ptr<TransactionStatistics> _transactionStatistics;
 };
 
 }  // namespace arangodb
