+++ tests/Mocks/StorageEngineMock.h
@@ -28,7 +28,6 @@
 #include "Futures/Future.h"
 #include "Metrics/IRegistry.h"
 #include "StorageEngine/HealthData.h"
-#include "Statistics/ServerStatistics.h"
 #include "StorageEngine/StorageEngine.h"
 #include "StorageEngine/TransactionState.h"
 #include "VocBase/Identifiers/IndexId.h"
@@ -223,10 +222,6 @@ class StorageEngineMock : public arangodb::StorageEngine {
   bool autoRefillIndexCaches() const override;
   bool autoRefillIndexCachesOnFollowers() const override;
 
-  arangodb::TransactionStatistics& transactionStatistics() noexcept override;
-  arangodb::TransactionStatistics const& transactionStatistics()
-      const noexcept override;
-
   static std::shared_ptr<arangodb::iresearch::IResearchLinkMock> buildLinkMock(
       arangodb::IndexId id, arangodb::LogicalCollection& collection,
       VPackSlice const& info);
@@ -257,5 +252,4 @@ class StorageEngineMock : public arangodb::StorageEngine {
   TRI_voc_tick_t _releasedTick;
   std::atomic_uint64_t _engineTick{100};
   arangodb::VersionTracker _versionTracker;
-  arangodb::TransactionStatistics _transactionStatistics;
 };
